pyproject.toml
[tool.poetry]
name = "sleep"
version = "0.1.0"
description = ""
authors = ["Noah Kantrowitz <noah@coderanger.net>", "C W <compiledwrong@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
withings-api = "^2.4.0"
django = "^5.0.2"
discord-py = "^2.2.3"
dj-database-url = "^2.0.0"
httpx = "^0.24.1"
uvicorn = "^0.22.0"
psycopg2 = "^2.9.6"
matplotlib = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
